+++ Packages/com.chisel.core/Chisel/Core/API.private/Managed/CSG/Jobs/FillVertexBuffersJob.cs
@@ -11,26 +11,27 @@ using Unity.Entities.UniversalDelegates;
 using Unity.Jobs;
 using Unity.Mathematics;
 using UnityEngine;
+using UnityEngine.Rendering;
 using ReadOnlyAttribute = Unity.Collections.ReadOnlyAttribute;
 
 namespace Chisel.Core
 {
     internal struct SubMeshCounts
     {
-        public MeshQuery meshQuery;
-        public int		surfaceParameter;
+        public MeshQuery    meshQuery;
+        public int		    surfaceParameter;
 
-        public int		meshQueryIndex;
-        public int		subMeshQueryIndex;
+        public int		    meshQueryIndex;
+        public int		    subMeshQueryIndex;
             
-        public uint	    geometryHashValue;  // used to detect changes in vertex positions  
-        public uint	    surfaceHashValue;   // used to detect changes in color, normal, tangent or uv (doesn't effect lighting)
+        public uint	        geometryHashValue;  // used to detect changes in vertex positions  
+        public uint	        surfaceHashValue;   // used to detect changes in color, normal, tangent or uv (doesn't effect lighting)
             
-        public int		vertexCount;
-        public int		indexCount;
+        public int		    vertexCount;
+        public int		    indexCount;
             
-        public int      surfacesOffset;
-        public int      surfacesCount;
+        public int          surfacesOffset;
+        public int          surfacesCount;
     };
 
     public struct SubMeshSection
@@ -100,6 +101,298 @@ namespace Chisel.Core
         }
     }
 
+    public struct SectionData
+    {
+        public int surfacesOffset;
+        public int surfacesCount;
+        public MeshQuery meshQuery;
+    }
+
+    internal struct SubMeshSurface
+    {
+        public int      brushNodeIndex;
+        public int      surfaceIndex;
+        public int      surfaceParameter;
+
+        public int      vertexCount;
+        public int      indexCount;
+
+        public uint     surfaceHash;
+        public uint     geometryHash;
+        public BlobAssetReference<ChiselBrushRenderBuffer> brushRenderBuffer;
+    }
+
+    [BurstCompile(CompileSynchronously = true)]
+    struct PrepareSubSectionsJob : IJobParallelFor
+    {
+        // Read
+        [NoAlias, ReadOnly] public NativeArray<MeshQuery>.ReadOnly  meshQueries;
+        [NoAlias, ReadOnly] public NativeArray<BrushData>           brushRenderData;
+
+        // Read, Write
+        [NativeDisableParallelForRestriction]
+        [NoAlias] public NativeListArray<SubMeshSurface>    subMeshSurfaces;
+
+        struct SubMeshSurfaceComparer : IComparer<SubMeshSurface>
+        {
+            public int Compare(SubMeshSurface x, SubMeshSurface y)
+            {
+                return x.surfaceParameter.CompareTo(y.surfaceParameter);
+            }
+        }
+
+        static readonly SubMeshSurfaceComparer subMeshSurfaceComparer = new SubMeshSurfaceComparer();
+
+        public void Execute()
+        {
+            for (int t = 0; t < meshQueries.Length; t++)
+            {
+                Execute(t);
+            }
+        }
+
+        public void Execute(int t)
+        {
+            int requiredSurfaceCount = 0;
+            for (int b = 0, count_b = brushRenderData.Length; b < count_b; b++)
+                requiredSurfaceCount += brushRenderData[b].brushSurfaceCount;
+
+            // THIS IS THE SLOWDOWN
+            // TODO: store surface separately from brushes, *somehow* make lifetime work
+            //              => multiple arrays, one for each meshQuery!
+            //              => sorted by surface.layerParameters[meshQuery.layerParameterIndex]!
+            //              => this whole job could be removed
+            // TODO: store surface info and its vertices/indices separately, both sequentially in arrays
+            // TODO: store surface vertices/indices sequentially in a big array, *somehow* make ordering work
+            // TODO: AllocateVertexBuffersJob/FillVertexBuffersJob and CopyToRenderMeshJob could be combined (one copy only)
+                
+            subMeshSurfaces.AllocateWithCapacityForIndex(t, requiredSurfaceCount);
+            var subMeshSurfaceList = subMeshSurfaces[t];
+
+            for (int b = 0, count_b = brushRenderData.Length; b < count_b; b++)
+            {
+                var brushData           = brushRenderData[b];
+                var brushRenderBuffer   = brushData.brushRenderBuffer;
+                ref var querySurfaces   = ref brushRenderBuffer.Value.querySurfaces[t]; // <-- 1. somehow this needs to 
+                                                                                        //     be in outer loop
+                ref var brushNodeIndex  = ref querySurfaces.brushNodeIndex;
+                ref var surfaces        = ref querySurfaces.surfaces;
+
+                for (int s = 0; s < surfaces.Length; s++) 
+                {
+                    subMeshSurfaceList.AddNoResize(new SubMeshSurface
+                    {
+                        brushNodeIndex      = brushNodeIndex,
+                        surfaceIndex        = surfaces[s].surfaceIndex,
+                        surfaceParameter    = surfaces[s].surfaceParameter, // <-- 2. store array per surfaceParameter => no sort
+                        vertexCount         = surfaces[s].vertexCount,
+                        indexCount          = surfaces[s].indexCount,
+                        surfaceHash         = surfaces[s].surfaceHash,
+                        geometryHash        = surfaces[s].geometryHash,
+                        brushRenderBuffer   = brushRenderBuffer, // <-- 3. Get rid of this somehow => memcpy
+                    });
+                }
+                // ^ do those 3 points (mentioned in comments)
+            }
+
+            subMeshSurfaceList.AsArray().Sort(subMeshSurfaceComparer);
+        }
+    }
+
+    [BurstCompile(CompileSynchronously = true)]
+    struct SortSurfacesParallelJob : IJob
+    {
+        const int kMaxPhysicsVertexCount = 64000;
+
+        // Read
+        [NoAlias, ReadOnly] public NativeArray<MeshQuery>.ReadOnly  meshQueries;
+        [NoAlias, ReadOnly] public NativeListArray<SubMeshSurface>  subMeshSurfaces;
+
+        // Write
+        [NoAlias, WriteOnly] public NativeList<SubMeshCounts> subMeshCounts;
+
+        // Per thread scratch memory
+        [NativeDisableContainerSafetyRestriction, NoAlias] NativeList<SubMeshCounts> sectionSubMeshCounts;
+
+        public void Execute()
+        {
+            // TODO: figure out why this order matters
+            for (int t = 0; t < subMeshSurfaces.Length; t++)
+            {
+                Execute(t);
+            }
+        }
+
+        public void Execute(int t)
+        {
+            var subMeshSurfaceList  = subMeshSurfaces[t];
+            var surfaceCount        = subMeshSurfaceList.Length;
+            if (surfaceCount == 0)
+                return;
+
+            var meshQuery       = meshQueries[t];
+            var isPhysics       = meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial;
+
+            if (sectionSubMeshCounts.IsCreated)
+            {
+                sectionSubMeshCounts.Clear();
+                if (sectionSubMeshCounts.Capacity < surfaceCount)
+                    sectionSubMeshCounts.Capacity = surfaceCount;
+            } else
+                sectionSubMeshCounts = new NativeList<SubMeshCounts>(surfaceCount, Allocator.Temp);
+
+            var currentSubMesh  = new SubMeshCounts
+            {
+                meshQueryIndex      = t,
+                subMeshQueryIndex   = 0,
+                meshQuery           = meshQuery,
+                surfacesOffset      = 0,
+                surfaceParameter    = subMeshSurfaceList[0].surfaceParameter
+            };
+            for (int b = 0; b < surfaceCount; b++)
+            {
+                var subMeshSurface              = subMeshSurfaceList[b];
+                var surfaceParameter            = subMeshSurface.surfaceParameter;
+                var surfaceVertexCount          = subMeshSurface.vertexCount;
+                var surfaceIndexCount           = subMeshSurface.indexCount;
+
+                if (currentSubMesh.surfaceParameter != surfaceParameter || 
+                    (isPhysics && currentSubMesh.vertexCount >= kMaxPhysicsVertexCount))
+                {
+                    // Store the previous subMeshCount
+                    if (currentSubMesh.indexCount > 0 && currentSubMesh.vertexCount > 0)
+                        sectionSubMeshCounts.AddNoResize(currentSubMesh);
+                        
+                    // Create the new SubMeshCount
+                    currentSubMesh.surfaceParameter     = surfaceParameter;
+                    currentSubMesh.subMeshQueryIndex++;
+                    currentSubMesh.surfaceHashValue     = 0;
+                    currentSubMesh.geometryHashValue    = 0;
+                    currentSubMesh.indexCount           = 0;
+                    currentSubMesh.vertexCount          = 0;
+                    currentSubMesh.surfacesOffset       += currentSubMesh.surfacesCount;
+                    currentSubMesh.surfacesCount        = 0;
+                } 
+
+                currentSubMesh.indexCount   += surfaceIndexCount;
+                currentSubMesh.vertexCount  += surfaceVertexCount;
+                currentSubMesh.surfaceHashValue  = math.hash(new uint2(currentSubMesh.surfaceHashValue, subMeshSurface.surfaceHash));
+                currentSubMesh.geometryHashValue = math.hash(new uint2(currentSubMesh.geometryHashValue, subMeshSurface.geometryHash));
+                currentSubMesh.surfacesCount++;
+            }
+            // Store the last subMeshCount
+            if (currentSubMesh.indexCount > 0 && currentSubMesh.vertexCount > 0)
+                sectionSubMeshCounts.AddNoResize(currentSubMesh);
+
+            subMeshCounts.AddRangeNoResize(sectionSubMeshCounts);
+        }
+    }
+
+    [BurstCompile(CompileSynchronously = true)]
+    struct GatherSurfacesJob : IJob
+    {
+        // Read / Write
+        [NoAlias] public NativeList<SubMeshCounts>              subMeshCounts;
+
+        // Write
+        [NoAlias, WriteOnly] public NativeList<SubMeshSection>  subMeshSections;
+            
+        public void Execute()
+        {
+            if (subMeshCounts.Length == 0)
+                return;
+
+            int descriptionIndex = 0;
+            //var contentsIndex = 0;
+            if (subMeshCounts[0].meshQuery.LayerParameterIndex == LayerParameterIndex.None ||
+                subMeshCounts[0].meshQuery.LayerParameterIndex == LayerParameterIndex.RenderMaterial)
+            {
+                var prevQuery = subMeshCounts[0].meshQuery;
+                var startIndex = 0;
+                for (; descriptionIndex < subMeshCounts.Length; descriptionIndex++)
+                {
+                    var subMeshCount = subMeshCounts[descriptionIndex];
+                    // Exit when layerParameterIndex is no longer LayerParameter1/None
+                    if (subMeshCount.meshQuery.LayerParameterIndex != LayerParameterIndex.None &&
+                        subMeshCount.meshQuery.LayerParameterIndex != LayerParameterIndex.RenderMaterial)
+                        break;
+
+                    var currQuery = subMeshCount.meshQuery;
+                    if (prevQuery == currQuery)
+                    {
+                        continue;
+                    }
+
+                    int totalVertexCount = 0;
+                    int totalIndexCount = 0;
+                    for (int i = startIndex; i < descriptionIndex; i++)
+                    {
+                        totalVertexCount += subMeshCounts[i].vertexCount;
+                        totalIndexCount += subMeshCounts[i].indexCount;
+                    }
+
+                    // Group by all subMeshCounts with same query
+                    subMeshSections.AddNoResize(new SubMeshSection
+                    {
+                        meshQuery           = subMeshCounts[startIndex].meshQuery,
+                        startIndex          = startIndex, 
+                        endIndex            = descriptionIndex,
+                        totalVertexCount    = totalVertexCount,
+                        totalIndexCount     = totalIndexCount,
+                    });
+
+                    startIndex = descriptionIndex;
+                    prevQuery = currQuery;
+                }
+
+                {
+                    int totalVertexCount = 0;
+                    int totalIndexCount = 0;
+                    for (int i = startIndex; i < descriptionIndex; i++)
+                    {
+                        totalVertexCount += subMeshCounts[i].vertexCount;
+                        totalIndexCount  += subMeshCounts[i].indexCount;
+                    }
+
+                    subMeshSections.AddNoResize(new SubMeshSection
+                    {
+                        meshQuery           = subMeshCounts[startIndex].meshQuery,
+                        startIndex          = startIndex,
+                        endIndex            = descriptionIndex,
+                        totalVertexCount    = totalVertexCount,
+                        totalIndexCount     = totalIndexCount
+                    });
+                }
+            }
+                
+
+            if (descriptionIndex < subMeshCounts.Length &&
+                subMeshCounts[descriptionIndex].meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial)
+            {
+                Debug.Assert(subMeshCounts[subMeshCounts.Length - 1].meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial);
+
+                // Loop through all subMeshCounts with LayerParameter2, and create collider meshes from them
+                for (int i = 0; descriptionIndex < subMeshCounts.Length; descriptionIndex++, i++)
+                {
+                    var subMeshCount = subMeshCounts[descriptionIndex];
+
+                    // Exit when layerParameterIndex is no longer LayerParameter2
+                    if (subMeshCount.meshQuery.LayerParameterIndex != LayerParameterIndex.PhysicsMaterial)
+                        break;
+
+                    subMeshSections.AddNoResize(new SubMeshSection
+                    {
+                        meshQuery           = subMeshCount.meshQuery,
+                        startIndex          = descriptionIndex,
+                        endIndex            = descriptionIndex,
+                        totalVertexCount    = subMeshCount.vertexCount,
+                        totalIndexCount     = subMeshCount.indexCount
+                    });
+                }
+            }
+        }
+    }
+
     [BurstCompile(CompileSynchronously = true)]
     struct AllocateVertexBuffersJob : IJob
     {
@@ -164,6 +457,41 @@ namespace Chisel.Core
         }
     }
 
+    [BurstCompile(CompileSynchronously = true)]
+    struct GenerateMeshDescriptionJob : IJob
+    {
+        [NoAlias, ReadOnly] public NativeArray<SubMeshCounts> subMeshCounts;
+
+        [NativeDisableParallelForRestriction]
+        [NoAlias] public NativeList<GeneratedMeshDescription> meshDescriptions;
+
+        public void Execute()
+        {
+            if (meshDescriptions.Capacity < subMeshCounts.Length)
+                meshDescriptions.Capacity = subMeshCounts.Length;
+
+            for (int i = 0; i < subMeshCounts.Length; i++)
+            {
+                var subMesh = subMeshCounts[i];
+
+                var description = new GeneratedMeshDescription
+                {
+                    meshQuery           = subMesh.meshQuery,
+                    surfaceParameter    = subMesh.surfaceParameter,
+                    meshQueryIndex      = subMesh.meshQueryIndex,
+                    subMeshQueryIndex   = subMesh.subMeshQueryIndex,
+
+                    geometryHashValue   = subMesh.geometryHashValue,
+                    surfaceHashValue    = subMesh.surfaceHashValue,
+
+                    vertexCount         = subMesh.vertexCount,
+                    indexCount          = subMesh.indexCount
+                };
+
+                meshDescriptions.Add(description);
+            }
+        }
+    }    
 
     [BurstCompile(CompileSynchronously = true)]
     struct FillVertexBuffersJob : IJobParallelFor
@@ -351,334 +679,268 @@ namespace Chisel.Core
             }
         }
     }
+    
 
-    [BurstCompile(CompileSynchronously = true)]
-    struct GenerateMeshDescriptionJob : IJob
+
+    public struct ChiselMeshUpdate
     {
-        [NoAlias, ReadOnly] public NativeArray<SubMeshCounts> subMeshCounts;
-
-        [NativeDisableParallelForRestriction]
-        [NoAlias] public NativeList<GeneratedMeshDescription> meshDescriptions;
-
-        public void Execute()
-        {
-            if (meshDescriptions.Capacity < subMeshCounts.Length)
-                meshDescriptions.Capacity = subMeshCounts.Length;
-
-            for (int i = 0; i < subMeshCounts.Length; i++)
-            {
-                var subMesh = subMeshCounts[i];
-
-                var description = new GeneratedMeshDescription
-                {
-                    meshQuery           = subMesh.meshQuery,
-                    surfaceParameter    = subMesh.surfaceParameter,
-                    meshQueryIndex      = subMesh.meshQueryIndex,
-                    subMeshQueryIndex   = subMesh.subMeshQueryIndex,
-
-                    geometryHashValue   = subMesh.geometryHashValue,
-                    surfaceHashValue    = subMesh.surfaceHashValue,
-
-                    vertexCount         = subMesh.vertexCount,
-                    indexCount          = subMesh.indexCount
-                };
-
-                meshDescriptions.Add(description);
-            }
-        }
+        public int contentsIndex;
+        public int meshIndex;
+        public int objectIndex;
     }
     
-    public struct SectionData
-    {
-        public int surfacesOffset;
-        public int surfacesCount;
-        public MeshQuery meshQuery;
-    }
-
-    internal struct SubMeshSurface
-    {
-        public int      brushNodeIndex;
-        public int      surfaceIndex;
-        public int      surfaceParameter;
-
-        public int      vertexCount;
-        public int      indexCount;
-
-        public uint     surfaceHash;
-        public uint     geometryHash;
-        public BlobAssetReference<ChiselBrushRenderBuffer> brushRenderBuffer;
-    }
-
     [BurstCompile(CompileSynchronously = true)]
-    struct PrepareSubSectionsJob : IJobParallelFor
+    public struct AssignMeshesJob : IJob
     {
-        // Read
-        [NoAlias, ReadOnly] public NativeArray<MeshQuery>.ReadOnly  meshQueries;
-        [NoAlias, ReadOnly] public NativeArray<BrushData>           brushRenderData;
-
-        // Read, Write
-        [NativeDisableParallelForRestriction]
-        [NoAlias] public NativeListArray<SubMeshSurface>    subMeshSurfaces;
-
-        struct SubMeshSurfaceComparer : IComparer<SubMeshSurface>
+        public const int kDebugHelperCount = 6;
+        public struct DebugRenderFlags { public LayerUsageFlags Item1; public LayerUsageFlags Item2; };
+        public static readonly DebugRenderFlags[] kGeneratedDebugRendererFlags = new DebugRenderFlags[kDebugHelperCount]
         {
-            public int Compare(SubMeshSurface x, SubMeshSurface y)
-            {
-                return x.surfaceParameter.CompareTo(y.surfaceParameter);
-            }
+            new DebugRenderFlags{ Item1 = LayerUsageFlags.None                  , Item2 = LayerUsageFlags.Renderable },              // is explicitly set to "not visible"
+            new DebugRenderFlags{ Item1 = LayerUsageFlags.RenderCastShadows     , Item2 = LayerUsageFlags.RenderCastShadows },       // casts Shadows and is renderered
+            new DebugRenderFlags{ Item1 = LayerUsageFlags.CastShadows           , Item2 = LayerUsageFlags.RenderCastShadows },       // casts Shadows and is NOT renderered (shadowOnly)
+            new DebugRenderFlags{ Item1 = LayerUsageFlags.RenderReceiveShadows  , Item2 = LayerUsageFlags.RenderReceiveShadows },    // any surface that receives shadows (must be rendered)
+            new DebugRenderFlags{ Item1 = LayerUsageFlags.Collidable            , Item2 = LayerUsageFlags.Collidable },              // collider surfaces
+            new DebugRenderFlags{ Item1 = LayerUsageFlags.Culled                , Item2 = LayerUsageFlags.Culled }                   // all surfaces removed by the CSG algorithm
+        };
+
+        [NoAlias, ReadOnly] public NativeList<GeneratedMeshDescription> meshDescriptions;
+        [NoAlias, ReadOnly] public NativeList<SubMeshSection>           subMeshSections;
+        [NoAlias, ReadOnly] public NativeList<Mesh.MeshData>            meshDatas;
+
+        [NoAlias, WriteOnly] public NativeList<Mesh.MeshData>        meshes;
+        [NoAlias, WriteOnly] public NativeList<ChiselMeshUpdate>     debugHelperMeshes;
+        [NoAlias, WriteOnly] public NativeList<ChiselMeshUpdate>     renderMeshes;
+
+        [NoAlias] public NativeList<ChiselMeshUpdate> colliderMeshUpdates;
+
+
+        [BurstDiscard]
+        public static void InvalidQuery(LayerUsageFlags query, LayerUsageFlags mask)
+        {
+            Debug.Assert(false, $"Invalid helper query used (query: {query}, mask: {mask})");
+
         }
 
-        static readonly SubMeshSurfaceComparer subMeshSurfaceComparer = new SubMeshSurfaceComparer();
-
-        public void Execute()
+        public void Execute() 
         {
-            for (int t = 0; t < meshQueries.Length; t++)
+            int meshIndex = 0;
+            int colliderCount = 0;
+            if (meshDescriptions.IsCreated)
             {
-                Execute(t);
-            }
-        }
-
-        public void Execute(int t)
-        {
-            int requiredSurfaceCount = 0;
-            for (int b = 0, count_b = brushRenderData.Length; b < count_b; b++)
-                requiredSurfaceCount += brushRenderData[b].brushSurfaceCount;
-
-            // THIS IS THE SLOWDOWN
-            // TODO: store surface separately from brushes, *somehow* make lifetime work
-            //              => multiple arrays, one for each meshQuery!
-            //              => sorted by surface.layerParameters[meshQuery.layerParameterIndex]!
-            //              => this whole job could be removed
-            // TODO: store surface info and its vertices/indices separately, both sequentially in arrays
-            // TODO: store surface vertices/indices sequentially in a big array, *somehow* make ordering work
-            // TODO: AllocateVertexBuffersJob/FillVertexBuffersJob and CopyToRenderMeshJob could be combined (one copy only)
-                
-            subMeshSurfaces.AllocateWithCapacityForIndex(t, requiredSurfaceCount);
-            var subMeshSurfaceList = subMeshSurfaces[t];
-
-            for (int b = 0, count_b = brushRenderData.Length; b < count_b; b++)
-            {
-                var brushData           = brushRenderData[b];
-                var brushRenderBuffer   = brushData.brushRenderBuffer;
-                ref var querySurfaces   = ref brushRenderBuffer.Value.querySurfaces[t]; // <-- 1. somehow this needs to 
-                                                                                        //     be in outer loop
-                ref var brushNodeIndex  = ref querySurfaces.brushNodeIndex;
-                ref var surfaces        = ref querySurfaces.surfaces;
-
-                for (int s = 0; s < surfaces.Length; s++) 
+                for (int i = 0; i < subMeshSections.Length; i++)
                 {
-                    subMeshSurfaceList.AddNoResize(new SubMeshSurface
+                    var subMeshSection = subMeshSections[i];
+                    if (subMeshSection.meshQuery.LayerParameterIndex == LayerParameterIndex.None)
                     {
-                        brushNodeIndex      = brushNodeIndex,
-                        surfaceIndex        = surfaces[s].surfaceIndex,
-                        surfaceParameter    = surfaces[s].surfaceParameter, // <-- 2. store array per surfaceParameter => no sort
-                        vertexCount         = surfaces[s].vertexCount,
-                        indexCount          = surfaces[s].indexCount,
-                        surfaceHash         = surfaces[s].surfaceHash,
-                        geometryHash        = surfaces[s].geometryHash,
-                        brushRenderBuffer   = brushRenderBuffer, // <-- 3. Get rid of this somehow => memcpy
-                    });
+                        int helperIndex = -1;
+                        var query   = subMeshSection.meshQuery.LayerQuery;
+                        var mask    = subMeshSection.meshQuery.LayerQueryMask;
+                        for (int f = 0; f < kGeneratedDebugRendererFlags.Length; f++)
+                        {
+                            if (kGeneratedDebugRendererFlags[f].Item1 != query ||
+                                kGeneratedDebugRendererFlags[f].Item2 != mask)
+                                continue;
+
+                            helperIndex = f;
+                            break;
+                        }
+                        if (helperIndex == -1)
+                        {
+                            InvalidQuery(query, mask);
+                            continue;
+                        }
+
+                        meshes.Add(meshDatas[meshIndex]);
+                        debugHelperMeshes.Add(new ChiselMeshUpdate
+                        {
+                            contentsIndex       = i,
+                            meshIndex           = meshIndex,
+                            objectIndex         = helperIndex
+                        });
+                        meshIndex++; 
+                    } else
+                    if (subMeshSection.meshQuery.LayerParameterIndex == LayerParameterIndex.RenderMaterial)
+                    {
+                        var renderIndex = (int)(subMeshSection.meshQuery.LayerQuery & LayerUsageFlags.RenderReceiveCastShadows);
+                        meshes.Add(meshDatas[meshIndex]);
+                        renderMeshes.Add(new ChiselMeshUpdate
+                        {
+                            contentsIndex       = i,
+                            meshIndex           = meshIndex,
+                            objectIndex         = renderIndex
+                        });
+                        meshIndex++;
+                    } else
+                    if (subMeshSection.meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial)
+                        colliderCount++;
                 }
-                // ^ do those 3 points (mentioned in comments)
             }
 
-            subMeshSurfaceList.AsArray().Sort(subMeshSurfaceComparer);
-        }
-    }
-
-    [BurstCompile(CompileSynchronously = true)]
-    struct SortSurfacesParallelJob : IJob
-    //struct SortSurfacesParallelJob : IJobParallelFor
-    {
-        const int kMaxPhysicsVertexCount = 64000;
-
-        // Read
-        [NoAlias, ReadOnly] public NativeArray<MeshQuery>.ReadOnly  meshQueries;
-        [NoAlias, ReadOnly] public NativeListArray<SubMeshSurface>  subMeshSurfaces;
-
-        // Write
-        [NoAlias, WriteOnly] public NativeList<SubMeshCounts>//.ParallelWriter <-- crashes when IJob
-                                        subMeshCounts;
-
-        // Per thread scratch memory
-        [NativeDisableContainerSafetyRestriction, NoAlias] NativeList<SubMeshCounts> sectionSubMeshCounts;
-
-        public void Execute()
-        {
-            // TODO: figure out why this order matters
-            for (int t = 0; t < subMeshSurfaces.Length; t++)
+            if (colliderMeshUpdates.Capacity < colliderCount)
+                colliderMeshUpdates.Capacity = colliderCount;
+            var colliderIndex = 0;
+            if (meshDescriptions.IsCreated)
             {
-                Execute(t);
-            }
-        }
-
-        public void Execute(int t)
-        {
-            var subMeshSurfaceList  = subMeshSurfaces[t];
-            var surfaceCount        = subMeshSurfaceList.Length;
-            if (surfaceCount == 0)
-                return;
-
-            var meshQuery       = meshQueries[t];
-            var isPhysics       = meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial;
-
-            if (sectionSubMeshCounts.IsCreated)
-            {
-                sectionSubMeshCounts.Clear();
-                if (sectionSubMeshCounts.Capacity < surfaceCount)
-                    sectionSubMeshCounts.Capacity = surfaceCount;
-            } else
-                sectionSubMeshCounts = new NativeList<SubMeshCounts>(surfaceCount, Allocator.Temp);
-
-            var currentSubMesh  = new SubMeshCounts
-            {
-                meshQueryIndex      = t,
-                subMeshQueryIndex   = 0,
-                meshQuery           = meshQuery,
-                surfacesOffset      = 0,
-                surfaceParameter    = subMeshSurfaceList[0].surfaceParameter
-            };
-            for (int b = 0; b < surfaceCount; b++)
-            {
-                var subMeshSurface              = subMeshSurfaceList[b];
-                var surfaceParameter            = subMeshSurface.surfaceParameter;
-                var surfaceVertexCount          = subMeshSurface.vertexCount;
-                var surfaceIndexCount           = subMeshSurface.indexCount;
-
-                if (currentSubMesh.surfaceParameter != surfaceParameter || 
-                    (isPhysics && currentSubMesh.vertexCount >= kMaxPhysicsVertexCount))
+                for (int i = 0; i < subMeshSections.Length; i++)
                 {
-                    // Store the previous subMeshCount
-                    if (currentSubMesh.indexCount > 0 && currentSubMesh.vertexCount > 0)
-                        sectionSubMeshCounts.AddNoResize(currentSubMesh);
-                        
-                    // Create the new SubMeshCount
-                    currentSubMesh.surfaceParameter     = surfaceParameter;
-                    currentSubMesh.subMeshQueryIndex++;
-                    currentSubMesh.surfaceHashValue     = 0;
-                    currentSubMesh.geometryHashValue    = 0;
-                    currentSubMesh.indexCount           = 0;
-                    currentSubMesh.vertexCount          = 0;
-                    currentSubMesh.surfacesOffset       += currentSubMesh.surfacesCount;
-                    currentSubMesh.surfacesCount        = 0;
-                } 
-
-                currentSubMesh.indexCount   += surfaceIndexCount;
-                currentSubMesh.vertexCount  += surfaceVertexCount;
-                currentSubMesh.surfaceHashValue  = math.hash(new uint2(currentSubMesh.surfaceHashValue, subMeshSurface.surfaceHash));
-                currentSubMesh.geometryHashValue = math.hash(new uint2(currentSubMesh.geometryHashValue, subMeshSurface.geometryHash));
-                currentSubMesh.surfacesCount++;
-            }
-            // Store the last subMeshCount
-            if (currentSubMesh.indexCount > 0 && currentSubMesh.vertexCount > 0)
-                sectionSubMeshCounts.AddNoResize(currentSubMesh);
-
-            subMeshCounts.AddRangeNoResize(sectionSubMeshCounts);
-        }
-    }
-
-    [BurstCompile(CompileSynchronously = true)]
-    struct GatherSurfacesJob : IJob
-    {
-        // Read / Write
-        [NoAlias] public NativeList<SubMeshCounts>              subMeshCounts;
-
-        // Write
-        [NoAlias, WriteOnly] public NativeList<SubMeshSection>  subMeshSections;
-            
-        public void Execute()
-        {
-            if (subMeshCounts.Length == 0)
-                return;
-
-            int descriptionIndex = 0;
-            //var contentsIndex = 0;
-            if (subMeshCounts[0].meshQuery.LayerParameterIndex == LayerParameterIndex.None ||
-                subMeshCounts[0].meshQuery.LayerParameterIndex == LayerParameterIndex.RenderMaterial)
-            {
-                var prevQuery = subMeshCounts[0].meshQuery;
-                var startIndex = 0;
-                for (; descriptionIndex < subMeshCounts.Length; descriptionIndex++)
-                {
-                    var subMeshCount = subMeshCounts[descriptionIndex];
-                    // Exit when layerParameterIndex is no longer LayerParameter1/None
-                    if (subMeshCount.meshQuery.LayerParameterIndex != LayerParameterIndex.None &&
-                        subMeshCount.meshQuery.LayerParameterIndex != LayerParameterIndex.RenderMaterial)
-                        break;
-
-                    var currQuery = subMeshCount.meshQuery;
-                    if (prevQuery == currQuery)
-                    {
+                    var subMeshSection = subMeshSections[i];
+                    if (subMeshSection.meshQuery.LayerParameterIndex != LayerParameterIndex.PhysicsMaterial)
                         continue;
-                    }
 
-                    int totalVertexCount = 0;
-                    int totalIndexCount = 0;
-                    for (int i = startIndex; i < descriptionIndex; i++)
+                    var surfaceParameter = meshDescriptions[subMeshSection.startIndex].surfaceParameter;
+
+                    meshes.Add(meshDatas[meshIndex]);
+                    colliderMeshUpdates.Add(new ChiselMeshUpdate
                     {
-                        totalVertexCount += subMeshCounts[i].vertexCount;
-                        totalIndexCount += subMeshCounts[i].indexCount;
-                    }
-
-                    // Group by all subMeshCounts with same query
-                    subMeshSections.AddNoResize(new SubMeshSection
-                    {
-                        meshQuery           = subMeshCounts[startIndex].meshQuery,
-                        startIndex          = startIndex, 
-                        endIndex            = descriptionIndex,
-                        totalVertexCount    = totalVertexCount,
-                        totalIndexCount     = totalIndexCount,
-                    });
-
-                    startIndex = descriptionIndex;
-                    prevQuery = currQuery;
+                        contentsIndex   = colliderIndex,
+                        meshIndex       = meshIndex,
+                        objectIndex     = surfaceParameter
+                    }); 
+                    colliderIndex++;
+                    meshIndex++;
                 }
+            }
+        }
+    }
 
-                {
-                    int totalVertexCount = 0;
-                    int totalIndexCount = 0;
-                    for (int i = startIndex; i < descriptionIndex; i++)
-                    {
-                        totalVertexCount += subMeshCounts[i].vertexCount;
-                        totalIndexCount  += subMeshCounts[i].indexCount;
-                    }
+    [BurstCompile(CompileSynchronously = true)]
+    public struct CopyToRenderMeshJob : IJobParallelFor
+    {
+        // Read
+        [NoAlias, ReadOnly] public NativeArray<VertexAttributeDescriptor>   renderDescriptors;
+        [NoAlias, ReadOnly] public NativeList<ChiselMeshUpdate>             renderMeshes;
+        [NoAlias, ReadOnly] public NativeListArray<GeneratedSubMesh>        subMeshes;
+        [NoAlias, ReadOnly] public NativeListArray<int> 	                indices;
+        [NoAlias, ReadOnly] public NativeListArray<RenderVertex>            vertices;
 
-                    subMeshSections.AddNoResize(new SubMeshSection
-                    {
-                        meshQuery           = subMeshCounts[startIndex].meshQuery,
-                        startIndex          = startIndex,
-                        endIndex            = descriptionIndex,
-                        totalVertexCount    = totalVertexCount,
-                        totalIndexCount     = totalIndexCount
-                    });
-                }
+        // Read (get meshData)/Write (write to meshData)
+        [NativeDisableContainerSafetyRestriction]
+        [NoAlias] public NativeList<Mesh.MeshData> meshes;
+
+        public void Execute(int renderIndex)
+        {
+            var update          = renderMeshes[renderIndex];
+            var contentsIndex   = update.contentsIndex;
+            var meshIndex       = update.meshIndex;
+            var meshData        = meshes[meshIndex];
+
+            if (!this.subMeshes.IsIndexCreated(contentsIndex) ||
+                !this.vertices.IsIndexCreated(contentsIndex) ||
+                !this.indices.IsIndexCreated(contentsIndex))
+            {
+                meshData.SetVertexBufferParams(0, renderDescriptors);
+                meshData.SetIndexBufferParams(0, IndexFormat.UInt32);
+                meshData.subMeshCount = 0;
+                return;
             }
                 
+            var subMeshesArray      = this.subMeshes[contentsIndex].AsArray();
+            var verticesArray       = this.vertices[contentsIndex].AsArray();
+            var indicesArray        = this.indices[contentsIndex].AsArray();
 
-            if (descriptionIndex < subMeshCounts.Length &&
-                subMeshCounts[descriptionIndex].meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial)
+
+            meshData.SetVertexBufferParams(verticesArray.Length, renderDescriptors);
+            meshData.SetIndexBufferParams(indicesArray.Length, IndexFormat.UInt32);
+
+            var dstVertices = meshData.GetVertexData<RenderVertex>(stream: 0);
+            dstVertices.CopyFrom(verticesArray);
+            
+            var dstIndices = meshData.GetIndexData<int>();
+            dstIndices.CopyFrom(indicesArray);
+
+            meshData.subMeshCount = subMeshesArray.Length;
+            for (int i = 0; i < subMeshesArray.Length; i++)
             {
-                Debug.Assert(subMeshCounts[subMeshCounts.Length - 1].meshQuery.LayerParameterIndex == LayerParameterIndex.PhysicsMaterial);
-
-                // Loop through all subMeshCounts with LayerParameter2, and create collider meshes from them
-                for (int i = 0; descriptionIndex < subMeshCounts.Length; descriptionIndex++, i++)
+                var srcBounds   = subMeshesArray[i].bounds;
+                var center      = (Vector3)((srcBounds.Max + srcBounds.Min) * 0.5f);
+                var size        = (Vector3)(srcBounds.Max - srcBounds.Min);
+                var dstBounds   = new Bounds(center, size);
+                meshData.SetSubMesh(i, new SubMeshDescriptor
                 {
-                    var subMeshCount = subMeshCounts[descriptionIndex];
-
-                    // Exit when layerParameterIndex is no longer LayerParameter2
-                    if (subMeshCount.meshQuery.LayerParameterIndex != LayerParameterIndex.PhysicsMaterial)
-                        break;
-
-                    subMeshSections.AddNoResize(new SubMeshSection
-                    {
-                        meshQuery           = subMeshCount.meshQuery,
-                        startIndex          = descriptionIndex,
-                        endIndex            = descriptionIndex,
-                        totalVertexCount    = subMeshCount.vertexCount,
-                        totalIndexCount     = subMeshCount.indexCount
-                    });
-                }
+                    baseVertex  = subMeshesArray[i].baseVertex,
+                    firstVertex = 0,
+                    vertexCount = subMeshesArray[i].vertexCount,
+                    indexStart  = subMeshesArray[i].baseIndex,
+                    indexCount  = subMeshesArray[i].indexCount,
+                    bounds      = dstBounds,
+                    topology    = UnityEngine.MeshTopology.Triangles,
+                }, MeshUpdateFlags.DontRecalculateBounds);
             }
+        }
+    }
+
+    [BurstCompile(CompileSynchronously = true)]
+    public struct CopyToColliderMeshJob : IJobParallelFor
+    {
+        // Read
+        [NoAlias, ReadOnly] public NativeArray<VertexAttributeDescriptor>   colliderDescriptors;
+        [NoAlias, ReadOnly] public NativeList<ChiselMeshUpdate>             colliderMeshes;
+        [NoAlias, ReadOnly] public NativeListArray<GeneratedSubMesh>        subMeshes;
+        [NoAlias, ReadOnly] public NativeListArray<int> 	                indices;
+        [NoAlias, ReadOnly] public NativeListArray<float3>                  vertices;
+        [NoAlias, ReadOnly] public int contentsIndex;
+        [NoAlias, ReadOnly] public int meshIndex;
+
+        // Read (get meshData)/Write (write to meshData)
+        [NativeDisableContainerSafetyRestriction]
+        [NoAlias] public NativeList<Mesh.MeshData> meshes;
+        
+        public void Execute(int colliderIndex)
+        {
+            var update          = colliderMeshes[colliderIndex];
+            var contentsIndex   = update.contentsIndex;
+            var meshIndex       = update.meshIndex;
+            var meshData        = meshes[meshIndex];
+
+            if (!this.subMeshes.IsIndexCreated(contentsIndex) ||
+                !this.vertices.IsIndexCreated(contentsIndex) ||
+                !this.indices.IsIndexCreated(contentsIndex))
+            {
+                meshData.SetVertexBufferParams(0, colliderDescriptors);
+                meshData.SetIndexBufferParams(0, IndexFormat.UInt32);
+                meshData.subMeshCount = 0;
+                return;
+            }
+                
+            var subMeshesArray  = this.subMeshes[contentsIndex].AsArray();
+            var verticesArray   = this.vertices[contentsIndex].AsArray();
+            var indicesArray    = this.indices[contentsIndex].AsArray();
+
+            meshData.SetVertexBufferParams(verticesArray.Length, colliderDescriptors);
+            meshData.SetIndexBufferParams(indicesArray.Length, IndexFormat.UInt32);
+
+            var dstVertices = meshData.GetVertexData<float3>(stream: 0);
+            dstVertices.CopyFrom(verticesArray);
+                
+            var dstIndices = meshData.GetIndexData<int>();
+            dstIndices.CopyFrom(indicesArray);
+                
+            meshData.subMeshCount = subMeshesArray.Length;
+            for (int i = 0; i < subMeshesArray.Length; i++)
+            {
+                var srcBounds   = subMeshesArray[i].bounds;
+                var center      = (Vector3)((srcBounds.Max + srcBounds.Min) * 0.5f);
+                var size        = (Vector3)(srcBounds.Max - srcBounds.Min);
+                var dstBounds   = new Bounds(center, size);
+
+                meshData.SetSubMesh(i, new SubMeshDescriptor
+                {
+                    baseVertex  = subMeshesArray[i].baseVertex,
+                    firstVertex = 0,
+                    vertexCount = subMeshesArray[i].vertexCount,
+                    indexStart  = subMeshesArray[i].baseIndex,
+                    indexCount  = subMeshesArray[i].indexCount,
+                    bounds      = dstBounds,
+                    topology    = UnityEngine.MeshTopology.Triangles,
+                }, MeshUpdateFlags.DontRecalculateBounds);
+            }
+
+            // TODO: Figure out why sometimes setting a mesh on a MeshCollider causes BakeMesh to be called by unity
+            //       (in which case this would happen serially on the main thread, which would be slower than calling it here)
+            //       And sometimes it's not called? (in which case calling BakeMesh here would be *slower*)
+            //       Also, if we use Unity.Physics then this wouldn't make sense at all
+            //Physics.BakeMesh(instanceID, false);
         }
     }
 }
